requirements.txt
astroid==2.8.3
autopep8==1.5.7
certifi==2018.8.24
chardet==3.0.4
cycler==0.11.0
defusedxml==0.5.0
Django==2.2.13
django-allauth==0.37.1
django-cors-headers==2.4.0
django-rest-auth==0.9.3
djangorestframework==3.9.1
fonttools==4.28.2
gunicorn==19.7.1
idna==2.7
isort==5.9.3
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
matplotlib==3.5.0
mccabe==0.6.1
numpy==1.21.4
oauthlib==2.1.0
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
platformdirs==2.4.0
prompt-toolkit==3.0.23
pycodestyle==2.8.0
Pygments==2.10.0
pylint==2.11.1
pyparsing==3.0.6
python-dateutil==2.8.2
python3-openid==3.1.0
pytz==2018.5
requests==2.21.0
requests-oauthlib==1.0.0
setuptools-scm==6.3.2
six==1.11.0
sqlparse==0.3.0
toml==0.10.2
tomli==1.2.2
typing-extensions==3.10.0.2
urllib3==1.24.2
wcwidth==0.2.5
whitenoise==3.3.1
wrapt==1.13.2
